# src/openai/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API documentation for the application
paths:
...
  /api/job/{id}:
    get:
      summary: Get job by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the job
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
  /api/jobrun/{id}:
    get:
      summary: Get job run by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the job run
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
...
  /api/jobruns/{id}:
    get:
      summary: List job runs by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the job run
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
...
  /api/jobruns/context/{id}:
    get:
      summary: Get job run context by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the job run context
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRunContext'

    post:
      summary: Set job run context by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the job run context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRunContext'
      responses:
        '201':
          description: Job run context set successfully
  /api/disposition/{id}:
    get:
      summary: Get disposition by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the disposition
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disposition'
  /api/disposition:
    post:
      summary: Create a new disposition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Disposition'
      responses:
        '201':
          description: Disposition created successfully
...
  /jobruns/delete/{id}:
    get:
      summary: Delete job run by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the job run
      responses:
        '204':
          description: Job run deleted successfully
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        prompt_id:
          type: string
          format: uuid
        status:
          type: string
        created_at:
          type: string
          format: date-time
    JobRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        status:
          type: string
        created_at:
          type: string
          format: date-time
    JobRunContext:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_run_id:
          type: string
          format: uuid
        context_data:
          type: string
        created_at:
          type: string
          format: date-time
    Disposition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time